refactor: 숫자 목록을 추출하는 로직을 NumbersExtractor로 분리
