+++ src/main/kotlin/stringaddcalculator/StringAddCalculator.kt
@@ -12,7 +12,7 @@ class StringAddCalculator {
             return text.toInt()
         }
 
-        val numbers = convertToNumbers(text)
+        val numbers = NumbersExtractor.extract(text)
         return numbers.sum()
     }
 
@@ -26,28 +26,8 @@ class StringAddCalculator {
         return text.matches(DIGITS_ONLY_REGEX)
     }
 
-    private fun convertToNumbers(text: String): List<Int> {
-        val strNumbers: List<String> = extractNumbers(text)
-        return strNumbers.map { strNumber ->
-            strNumber.toInt()
-        }
-    }
-
-    private fun extractNumbers(text: String): List<String> {
-        return if (text.contains(CUSTOM_DELIMITER_REGEX)) {
-            val matchResult = CUSTOM_DELIMITER_REGEX.find(text) ?: return emptyList()
-
-            val delimiter = matchResult.groupValues[1]
-            matchResult.groupValues[2].split(delimiter)
-        } else {
-            text.split(DEFAULT_DELIMITER_REGEX)
-        }
-    }
-
     companion object {
         private val NEGATIVE_NUMBER_REGEX = Regex("-[0-9]+\$")
         private val DIGITS_ONLY_REGEX = Regex("^[0-9]+\$")
-        private val DEFAULT_DELIMITER_REGEX = Regex("[,:]")
-        private val CUSTOM_DELIMITER_REGEX = Regex("//(.)\\n(.*)")
     }
 }
